chore(refactor): split out dispatcher periodic code
Just clean up the dispatcher file, splitting
out all the periodic checks etc into their
own file, leaving the bones of dispatcher in one
place
